express-jwt added
Using express-jwt wrapper for more convenient approach to token handling
