+++ backend-api/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
+    "express-jwt": "^5.3.1",
     "google-auth-library": "^2.0.0",
     "jsonwebtoken": "^8.3.0",
     "morgan": "^1.9.1"
