+++ backend-api/src/controllers/auth.js
@@ -1,14 +1,47 @@
-const Authentication = require('../auth');
-const auth = new Authentication();
+const jwt = require('../auth/jwt');
+const googleAuth = require('../auth/google-auth');
 const InMemoryUserStore = require('../in-memory-user-store')
 
 const userStore = new InMemoryUserStore();
 
 
-exports.login = async (req, res) => {
+const createToken = user => {
+    return jwt.generateToken(user);
+  };
+  
+const getUser = login => {
+    const type = login.type;
+    switch (type) {
+      case 'google':
+        return googleAuth
+          .getGoogleUser(login.code)
+          .then(response => {
+            const content = {
+              token: createToken(response),
+              user: response
+            };
+            return content;
+          })
+          .catch(e => {
+            console.log(e);
+            throw new Error(e);
+          });
+        break;
+      default:
+        throw new Error('unknow token type [' + type + ']');
+    }
+  };
+
+const authenticate = login => {
+    return getUser(login).then(principal => {
+      return principal;
+    });
+  };
+
+const login = async (req, res) => {
     try {
         const login = req.body;
-        auth.authenticate(login).then(credentials => {
+        authenticate(login).then(credentials => {
         userStore.push(credentials.user);
         res.json({ success: true, data: credentials }).end();
       });
@@ -17,3 +50,5 @@ exports.login = async (req, res) => {
     } finally {
     }
 };
+
+module.exports.login = login;
