+++ backend-api/src/routes/auth.js
@@ -1,5 +1,5 @@
 const Auth = require('../controllers/auth');
-const {wrapAsync} = require('../utils/controller');
+const {wrapAsync} = require('../utils/controllers');
 
 module.exports = api => {
 	api.route('/auth/login').post(wrapAsync(Auth.login));
