+++ backend-api/src/routes/test.js
@@ -1,5 +1,5 @@
 const Test = require('../controllers/test');
-const {wrapAsync} = require('../utils/controller');
+const {wrapAsync} = require('../utils/controllers');
 
 module.exports = api => {
     api.route('/ping/public').get(wrapAsync(Test.pingPublic));
